requirements.txt
humanize~=3.11.0
Pillow~=5.4.1
numpy~=1.16.2
posix_ipc~=1.0.5
caldav~=0.8.0
pytz~=2021.1
icalevents==0.1.24
requests~=2.21.0
distro~=1.6.0
python_weather~=0.3.6
evdev~=1.4.0
PySensors~=0.0.4
htmlwebshot~=0.1.2
DateTime~=4.3
psutil~=5.5.1
asyncio~=3.4.3
httplib2~=0.19.1
urllib3~=1.24.1
